+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # alpha-test-app
 
+## 0.1.0-alpha.83
+
+### Patch Changes
+
+- Updated dependencies [5e96001]
+  - sit-onyx@1.0.0-alpha.86
+
+## 0.1.0-alpha.82
+
+### Patch Changes
+
+- Updated dependencies [641dac7]
+- Updated dependencies [641dac7]
+- Updated dependencies [afe16cf]
+  - sit-onyx@1.0.0-alpha.85
+
 ## 0.1.0-alpha.81
 
 ### Patch Changes