+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,21 @@
 # playground
 
+## 0.0.1-alpha.25
+
+### Patch Changes
+
+- Updated dependencies [5e96001]
+  - sit-onyx@1.0.0-alpha.86
+
+## 0.0.1-alpha.24
+
+### Patch Changes
+
+- Updated dependencies [641dac7]
+- Updated dependencies [641dac7]
+- Updated dependencies [afe16cf]
+  - sit-onyx@1.0.0-alpha.85
+
 ## 0.0.1-alpha.23
 
 ### Patch Changes