+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-alpha.86
+
+### Patch Changes
+
+- Updated dependencies [5e96001]
+  - sit-onyx@1.0.0-alpha.86
+
+## 1.0.0-alpha.85
+
+### Patch Changes
+
+- Updated dependencies [641dac7]
+- Updated dependencies [641dac7]
+- Updated dependencies [afe16cf]
+  - sit-onyx@1.0.0-alpha.85
+
 ## 1.0.0-alpha.84
 
 ### Patch Changes